code comments and cleanup
